feat: add dashboard routes and controller for summary and chart data
