+++ src/controllers/barangKeluarController.js
@@ -75,65 +75,6 @@ const barangKeluarController = {
         message: error.message || 'Internal server error'
       });
     }
-  },
-
-  // Get chart data for barang keluar
-  getChartData: async (req, res) => {
-    try {
-      const { startDate, endDate } = req.query;
-
-      const barangKeluarData = await BarangKeluar.findAll({
-        where: {
-          tanggal: {
-            [Op.between]: [
-              startDate || new Date(new Date().setDate(new Date().getDate() - 30)),
-              endDate || new Date()
-            ]
-          }
-        },
-        include: [{ model: Barang, attributes: ['nama_barang'] }],
-        attributes: [
-          [sequelize.fn('DATE', sequelize.col('tanggal')), 'date'],
-          [sequelize.fn('SUM', sequelize.col('jumlah')), 'total']
-        ],        group: [sequelize.fn('DATE', sequelize.col('tanggal')), 'Barang.id_barang', 'Barang.nama_barang'],
-        order: [[sequelize.fn('DATE', sequelize.col('tanggal')), 'ASC']]
-      });
-
-      // Format data untuk chart
-      const formattedData = barangKeluarData.reduce((acc, item) => {
-        const date = item.get('date');
-        if (!acc[date]) {
-          acc[date] = {
-            date,
-            total: 0
-          };
-        }
-        acc[date].total += parseInt(item.get('total'));
-        return acc;
-      }, {});
-
-      const chartData = Object.values(formattedData);
-
-      res.json({
-        success: true,
-        data: {
-          labels: chartData.map(item => item.date),
-          datasets: [{
-            label: 'Jumlah Barang Keluar per Hari',
-            data: chartData.map(item => item.total),
-            fill: false,
-            borderColor: '#FF6384',
-            tension: 0.1
-          }]
-        }
-      });
-    } catch (error) {
-      console.error('Error:', error);
-      res.status(500).json({
-        success: false,
-        message: error.message || 'Internal server error'
-      });
-    }
   }
 };
