+++ src/routes/barangMasuk.js
@@ -5,6 +5,5 @@ const auth = require('../middleware/auth');
 
 router.get('/', auth, barangMasukController.getAllBarangMasuk);
 router.post('/', auth, barangMasukController.createBarangMasuk);
-router.get('/chart', auth, barangMasukController.getChartData);
 
 module.exports = router;
